[WPE] `TestUIClient` `/webkit/WebKitWebView/geolocation-permission-requests` API test times out https://bugs.webkit.org/show_bug.cgi?id=261199
Unreviewed test gardening.

For this test to finish one of the
`navigator.geolocation.getCurrentPosition()` callbacks need to be
called, either the success or the error.

When we don't have access to `org.freedesktop.GeoClue2` DBus service,
the error callback will be called quite soon.

When we do have access, the time needed to get the current position can
exceed the default 5 seconds we give for the test to finish.

This test times out only on WPE Release bot. This is the only Linux bot
running on bare hardware, not in a container. It seems like the
container environment doesn't provide access to the GeoClue DBus server.
That explains why we don't get the timeout on other Linux bots.

Tools/TestWebKitAPI/glib/TestExpectations.json:

Canonical link: https://commits.webkit.org/267736@main