[Xcode] Track output files for symlinks created by WebKit script phases https://bugs.webkit.org/show_bug.cgi?id=261247 rdar://115085679
Reviewed by Alexey Proskuryakov.

Symlinks to daemons and XPC services do not need to get recreated when
their targets are rebuilt. So, it's fine to have the script phases which
create them declare them as outputs with no inputs.

Merge the "Add XPCServices symlink" and "Create symlinks to XPC
services" build phases. On Mac, the former creates
WebKit.framework/XPCServices and the latter creates symlinks in
WebKit.framework/Versions/A/XPCServices. But on other platforms, they
indicate that they *both* create WebKit.framework/XPCServices. Since
nothing in the build actually depends on the top-level XPCServices
symlink (iirc, it's for dyld), just create it as a side effect of the
"Create symlinks to XPC services" phase on Mac.

While they never get created in production-style builds, the outputs
need to use $(BUILT_PRODUCTS_DIR) and not $(TARGET_BUILD_DIR) to avoid
overlapping with the actual paths XPC services are installed to.

* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/267813@main