[visionOS] Audio does not spatialize correctly with multiple windows after switching tabs https://bugs.webkit.org/show_bug.cgi?id=261367 <radar://114060695>
Reviewed by NOBODY (OOPS!).

When we made spatial audio work for backgrounded tabs playing audio it broke the multi-window
spatial audio since the identifier is attached to the scene and not the window.

It would seem preferable to use an identifier attached to the window, but until we have such an
identifier, we can fix the regression by adapting the SPI to reset the behavior so AVSampleBufferRenderSynchronizer
clients can re-elect to use the renderer tracked by STS.

* Source/WebCore/PAL/pal/spi/cocoa/AVFoundationSPI.h:
Stage SPI.

* Source/WebCore/platform/graphics/avfoundation/objc/MediaPlayerPrivateMediaSourceAVFObjC.mm:
(WebCore::MediaPlayerPrivateMediaSourceAVFObjC::setPageIsVisible):
Reset the renderer after setting the identifier to nil.