WebGLRenderingContextBase::create checks for webGLEnabled setting unneccessarily https://bugs.webkit.org/show_bug.cgi?id=261398 rdar://problem/115273142
Reviewed by Antti Koivisto.

Remove the checks, the check is done in the caller.

For offscreen canvas, one of the checks would ref nullptr if the caller
did not do the check.

It is not right to send the webglcontextcreationerror if the setting is
disabled. When the setting is disabled, it should behave as if WebGL
was not implemented.

* Source/WebCore/html/canvas/WebGLRenderingContextBase.cpp:
(WebCore::WebGLRenderingContextBase::create):

Canonical link: https://commits.webkit.org/267859@main