visionOS: backdrop-filter in a transparent web view on top of material background renders black https://bugs.webkit.org/show_bug.cgi?id=261419 rdar://112798317
Reviewed by Mike Wyrzykowski.

* Source/WebCore/platform/graphics/ca/cocoa/PlatformCAFiltersCocoa.mm:
(WebCore::PlatformCAFilters::setFiltersOnLayer):
Use the visionOS-specific `normalizeEdgesTransparent` filter option in order
to make backdrops render correctly on top of material backgrounds.

Technically we only need this in cases where we're on top of a material, and could
use the cheaper `normalizeEdges` in the opaque-web-view case, but that would
require plumbing that knowledge through the fairly-distant filter application code.

Canonical link: https://commits.webkit.org/267875@main