Always install LayoutTestHelper at the same path as WebKitTestRunner https://bugs.webkit.org/show_bug.cgi?id=261425 rdar://115195681
Reviewed by Elliott Williams.

If we install at a different path, we can end up with a mostly empty WebKit.framework in
one of the locations, only containing LayoutTestHelper.

WebKit framework is not necessarily the best location for the tools, and we may want to
find a different one, but for now, making the minimal change to fix the issue.

* Tools/DumpRenderTree/mac/Configurations/LayoutTestHelper.xcconfig:

Canonical link: https://commits.webkit.org/267889@main