OffscreenCanvas WebGL does not produce console log messages https://bugs.webkit.org/show_bug.cgi?id=261400 rdar://115275708
Reviewed by Dan Glastonbury.

Use ScriptExecutionContext of the CanvasBase to add the console message.
This works for HTMLElement as well as OffscreenCanvas.

Do not reset the error counter on context restore. The counter is logical
to be for the context object, not for the underlying OpenGL context.

* Source/WebCore/html/canvas/WebGLRenderingContextBase.cpp:
(WebCore::WebGLRenderingContextBase::WebGLRenderingContextBase):
(WebCore::WebGLRenderingContextBase::initializeContextState):
(WebCore::WebGLRenderingContextBase::shouldPrintToConsole const):
(WebCore::WebGLRenderingContextBase::printToConsole):
* Source/WebCore/html/canvas/WebGLRenderingContextBase.h:

Canonical link: https://commits.webkit.org/267895@main