Crash in Vector::expandCapacity; WebCore::ISOFairPlayStreamingKeyContextBox::parse() https://bugs.webkit.org/show_bug.cgi?id=258712 rdar://111171940
Reviewed by Youenn Fablet.

In a number of cases, malformed or malicious boxes claim to have a size much larger
than the memory passed to the parser. Cap the m_size parameter to the remaining size
of the memory buffer to avoid situations where we attempt to create a vector big
enough to hold MAX_INT entries.

* Source/WebCore/platform/graphics/iso/ISOBox.cpp:
(WebCore::ISOBox::peekBox):
(WebCore::ISOBox::parse):

Originally-landed-as: 265870.2@safari-7616-branch (dfdca2cf04a5). rdar://116423096