[WGSL] Fix source locations https://bugs.webkit.org/show_bug.cgi?id=261521 rdar://115438296
Reviewed by Dan Glastonbury.

In 267855@main I implemented template disambiguation, which required changing how
we handle tokens: instead of consuming them progressively as we parse, we now need
to generate a list of tokens upfront. However, the parser was still using the lexer's
current position in order to assign source locations to AST nodes, but since all the
lexing was done upfront the current location was always the end of the file. In order
to fix it we keep track of the current location in the parser itself, instead of
consulting the lexer.

* Source/WebGPU/WGSL/Lexer.h:
(WGSL::Lexer::currentPosition const): Deleted.
* Source/WebGPU/WGSL/Parser.cpp:
(WGSL::Parser<Lexer>::consume):
(WGSL::Parser<Lexer>::parsePostfixExpression):
* Source/WebGPU/WGSL/ParserPrivate.h:

Canonical link: https://commits.webkit.org/267985@main