Add a test enumerating all the possible behaviors of the various context2d backends https://bugs.webkit.org/show_bug.cgi?id=261505 rdar://115425330
Reviewed by Matt Woodrow.

Add a test enumerating various combinations of context2d backend
creation parameters. This way it is easier to check the effect of
a change to parameter limits or other logic.

Notable result is that some cases where Accelerated is used will not
result in an actual Accelerated ImageBuffer. This will be addressed
in later commits.

The test creates normal context2d instances with various sizes and
lists out the actual backend. This way when the backend limits are
changed, the changes are reflected in the results.

The test creates context2d instances with forced ImageBuffer rendering
modes. This all the backend behaviors with various sizes are enumerated.

* LayoutTests/fast/canvas/image-buffer-backend-variants-expected.txt: Added.
* LayoutTests/fast/canvas/image-buffer-backend-variants.html: Added.
* LayoutTests/fast/canvas/image-buffer-iosurface-disabled-expected.txt: Removed.
* LayoutTests/fast/canvas/image-buffer-iosurface-disabled.html: Removed.
* Source/WebCore/html/canvas/CanvasRenderingContext2D.idl:
* Source/WebCore/html/canvas/CanvasRenderingContext2DBase.cpp:
(WebCore::CanvasRenderingContext2DBase::adjustImageBufferOptionsForTesting):
(WebCore::CanvasRenderingContext2DBase::getEffectiveRenderingModeForTesting):
* Source/WebCore/html/canvas/CanvasRenderingContext2DBase.h:
* Source/WebCore/html/canvas/CanvasRenderingContext2DSettings.h:
* Source/WebCore/html/canvas/CanvasRenderingContext2DSettings.idl:

Canonical link: https://commits.webkit.org/268015@main