[Filters] Referenced SVG filters with outsets get clipped https://bugs.webkit.org/show_bug.cgi?id=261615 rdar://115567857
Reviewed by NOBODY (OOPS!).

When updating the CSSFilter filterRegion, we should update the filterRegion of
all referenced SVGFilters.

Because the CSSFilter should not clip its effects, the outsets of the referenced
SVG filter should take into account the outsets of the SVGFilterElement itself.
These outsets are used to calculate the Filter maxEffectRect. The maxEffectRect
is the default primitive subregion.

* LayoutTests/css3/filters/effect-reference-outsets-expected.html: Added.
* LayoutTests/css3/filters/effect-reference-outsets.html: Added.
Add new tests.

* LayoutTests/css3/filters/reference-filter-outsets-expected.html:
* LayoutTests/css3/filters/svg-morphology-clipped-expected.html:
The expected files for these tests are incorrect.

* Source/WebCore/platform/RectEdges.h:
(WebCore::unionEdges):
* Source/WebCore/rendering/CSSFilter.cpp:
(WebCore::calculateReferenceFilterOutsets):
(WebCore::CSSFilter::setFilterRegion):