[NEW TEST] [ macOS ] API tests brings up a "There is no application set to open the URL callback:playing." dialogue box https://bugs.webkit.org/show_bug.cgi?id=261529 rdar://115450715
Reviewed by Eric Carlson.

This patch fixes the issue that was causing a dialog box to appear when running
TestWebKitAPI.WKWebViewSuspendAllMediaPlayback.PauseWhenResume. When the test
Loads video-with-audio.html, the javascript posts the message "playing", which
triggers an exception because the test does not provide a completion handler for
this message (because a completion handler is not needed). The code within the
catch block causes the dialog box to appear. To fix this, the catch block is now
empty, which unifies the behavior with the other javascript functions in the file.

* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKWebViewSuspendAllMediaPlayback.mm:
(TEST): un-disable test
* Tools/TestWebKitAPI/Tests/WebKitLegacy/ios/video-with-audio.html: Make catch block
In playing() empty

Canonical link: https://commits.webkit.org/268036@main